Refactor: レビュー指摘対応（招待コード機能の改善）
- groupInviteLinksService: 無意味な try/catch を削除しシンプル化
- groupInviteLinksService: JSX を含まないため拡張子を .tsx → .ts に変更
- groups/join/page: 招待コードをトリムしてからAPIに渡し、空白入りでも有効化
- groups/share-invite/page: buildShareMessage の重複呼び出しを変数化
